Optionally allow specifying agent environments
This allows you to specify a list of environments for each agent
profile. If set, the agent will autoregister with all specified
environments, not just the one the job was launched as.

Additionally, the scheduler will not launch new Aurora jobs to handle a
given GoCD job if the GoCD job is not in one of the allowed environments
for the requested cluster profile.

Also adds some very simple testing for agent profile validation.